Reorganise acceptance account context into account and lpa
